requirements.txt
django
pandas
mysqlclient
django_session_timeout
pillow
tensorflow
# deepface 
# dlib
# cmake
